+++ components/Attendance/Attendance.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './attendance.module.css';
+
+export const Attendance: React.FC = () => {
+    return (
+        <div>
+            Attendance
+        </div>
+    )
+}
