+++ components/ClassView/ClassViewMenu.tsx
@@ -5,11 +5,11 @@ import styles from "./ClassView.module.css";
 const BackButton: React.FC = () => {
     const router = useRouter();
     const onClick = () => {
-        router.back();
+        router.push("/home");
     }
     return (
         <>
-            <img src="back_button.png" onClick={onClick} alt="back" className={styles.backbutton}/>
+            <img src="back_button.svg" onClick={onClick} alt="back" className={styles.backbutton}/>
         </>
     )
 }
@@ -19,15 +19,18 @@ export const ClassViewMenu: React.FC = () => {
     const id = router.query.classid;
 
     return (
-    <>
-      <BackButton/>
       <nav className={styles.navbar}>
             <ul className={styles.navmenu}>
+            <BackButton/>
+              <li className={styles.navtitle}>
+                Java 1-2: Cat
+              </li>
               <li className={styles.navitem}>
                 <a
                   // href="/attendance"
                   className={router.pathname.includes("attendance") ? styles.clicked : styles.navlink}
-                  onClick={() => { router.push("/class/" + id + "/attendance") }}
+                  // onClick={() => { router.push("/class/" + id + "/attendance") }}
+                  onClick={() => { router.push("/class/attendance") }}
                 >
                   Attendance
                 </a>
@@ -37,7 +40,8 @@ export const ClassViewMenu: React.FC = () => {
                 <a
                   id="roster"
                   className={router.pathname.includes("roster") ? styles.clicked : styles.navlink}
-                  onClick={ () => { router.push("/class/" + id + "/roster") }}
+                  // onClick={ () => { router.push("/class/" + id + "/roster") }}
+                  onClick={ () => { router.push("/class/roster") }}
                 >
                   Roster
                 </a>
@@ -46,13 +50,13 @@ export const ClassViewMenu: React.FC = () => {
                 <a
                   id="modules"
                   className={router.pathname.includes("modules") ? styles.clicked : styles.navlink}
-                  onClick={ () => { router.push("/class/" + id + "/modules") }}
+                  // onClick={ () => { router.push("/class/" + id + "/modules") }}
+                  onClick={ () => { router.push("/class/modules") }}
                 >
                   Modules
                 </a>
               </li>
             </ul>
       </nav>
-    </>
     )
 }
