+++ components/ModuleView/Module.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './module.module.css';
+
+export const Module: React.FC = () => {
+    return (
+        <div>
+            Module
+        </div>
+    )
+}
