--- pages/class/[classid]/modules.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { ClassViewMenu } from '../../../components/ClassView/ClassViewMenu';
-import type { NextApplicationPage } from "../../_app";
-
-const ModulesPage: NextApplicationPage = () => {
-    return (
-        <div>
-            <ClassViewMenu/>
-            Modules
-        </div>
-    )
-  } 
-
-  export default ModulesPage;
