+++ pages/class/index.tsx
@@ -1,9 +1,10 @@
 import React, { useContext } from "react";
 import type { NextApplicationPage } from "../_app";
 import { AuthContext } from "../../context/AuthContext";
-
+import styles from "../../styles/class.module.css";
 import { Error } from "../../components/util/Error";
 import { ClassViewMenu } from "../../components/ClassView/ClassViewMenu";
+import { Roster } from "../../components/Roster/Roster";
 
 //This is the page that is rendered when the 'Home' button from the Navbar is clicked
 const Class: NextApplicationPage = () => {
@@ -12,7 +13,10 @@ const Class: NextApplicationPage = () => {
   if (user == null) return <Error />;
 
   return (
-    <ClassViewMenu/>
+    <div className={styles.container}>
+            <ClassViewMenu/>
+            <Roster/>
+        </div>
   );
 };
