+++ styles/class.module.css
@@ -1,0 +1,7 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    /* background-color: aquamarine; */
+    /* height: 100%; */
+}
